requirements.txt
Django==5.0.7					# BSD
django-extensions==3.2.3		# MIT
djangorestframework==3.15.2		# BSD - like
mysqlclient==2.2.4              # GPL2
whitenoise==6.7.0				# MIT
ijson==3.3.0					# BSD - like
questionary==2.0.1				# MIT
geopy==2.4.1					# MIT
scipy==1.14.0					# BSD - like
coloredlogs==15.0.1				# MIT
py7zr==0.20.8					# LGPL
pytz==2024.1					# MIT
requests==2.32.3				# Apache
pyroutelib3==1.7.2				# GPL3 (but this is only used by WarsawGTFS, not by us)
